requirements.txt
alembic==1.16.2
attrs==25.3.0
blinker==1.7.0
Brotli==1.1.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cryptography==45.0.5
Flask==2.3.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.3
h11==0.16.0
h2==4.2.0
hpack==4.1.0
hyperframe==6.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kaitaistruct==0.10
Mako==1.3.10
MarkupSafe==3.0.2
outcome==1.3.0.post0
packaging==25.0
pyasn1==0.6.1
pycparser==2.22
pydivert==2.1.0
PyMySQL==1.1.1
pyOpenSSL==25.1.0
pyparsing==3.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
referencing==0.36.2
requests==2.32.4
rpds-py==0.26.0
selenium==4.34.1
selenium-wire==5.1.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.41
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.14.1
urllib3==2.5.0
webdriver-manager==4.0.2
websocket-client==1.8.0
Werkzeug==3.1.3
wheel==0.45.1
wsproto==1.2.0
zstandard==0.23.0
